+++ components/Home/HashtagContents.tsx
@@ -12,19 +12,21 @@ type Props = {
 const HashtagContents: React.FC<Props> = ({ data }) => {
   return (
     <HashtagContentsLayout>
-      <SectionTitle>#랜덤글</SectionTitle>
-      <ImageContentsContainer>
-        {data.map((content, index) => {
-          return (
-            <ImageContent
-              key={index}
-              src={content.image}
-              alt={'date'}
-              postId={content.postId}
-            />
-          );
-        })}
-      </ImageContentsContainer>
+      <div className="center">
+        <SectionTitle>#랜덤글</SectionTitle>
+        <ImageContentsContainer>
+          {data.map((content, index) => {
+            return (
+              <ImageContent
+                key={index}
+                src={content.image}
+                alt={'date'}
+                postId={content.postId}
+              />
+            );
+          })}
+        </ImageContentsContainer>
+      </div>
     </HashtagContentsLayout>
   );
 };
