+++ styles/imageContainer.ts
@@ -6,16 +6,16 @@ export const ImageContentsContainer = styled.div`
   width: 100%;
   margin: 0 auto;
   display: grid;
-  grid-template-columns: repeat(4, 1fr);
+  grid-template-columns: repeat(2, 1fr);
   place-items: center;
   justify-content: center;
   align-items: center;
   grid-gap: 4px;
-  @media screen and (max-width: 1260px) {
+  @media screen and (min-width: 460px) {
     grid-template-columns: repeat(3, 1fr);
   }
-  @media screen and (max-width: 768px) {
-    grid-template-columns: repeat(2, 1fr);
+  @media screen and (min-width: 768px) {
+    grid-template-columns: repeat(4, 1fr);
   }
 `;
